requirements.txt
discord.py
steamio
donphan
discord-ext-menus
discord-ext-alternatives
discord-ext-ipc
humanize
matplotlib
jishaku
typer
asyncpg==0.22